+++ src/main/java/im/aop/loggers/messageinterpolation/ReflectionToStringStrategy.java
@@ -1,24 +1,41 @@
 package im.aop.loggers.messageinterpolation;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Objects;
 
-import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.builder.ReflectionToStringBuilder;
 import org.apache.commons.lang3.builder.ToStringStyle;
+import org.springframework.boot.context.properties.source.InvalidConfigurationPropertyValueException;
 
 public class ReflectionToStringStrategy implements ToStringStrategy {
 
   private final ReflectionToStringProperties reflectionToStringProperties;
 
+  private final List<Class<?>> supportedBaseClasses = new ArrayList<Class<?>>();
+
   public ReflectionToStringStrategy(
       final ReflectionToStringProperties reflectionToStringProperties) {
     this.reflectionToStringProperties = Objects.requireNonNull(reflectionToStringProperties);
+
+    for (String baseClass : reflectionToStringProperties.getBaseClasses()) {
+      try {
+        supportedBaseClasses.add(Class.forName(baseClass));
+      } catch (ClassNotFoundException e) {
+        throw new InvalidConfigurationPropertyValueException(
+            ReflectionToStringProperties.PREFIX + ".base-classes", baseClass, e.getMessage());
+      }
+    }
   }
 
   @Override
   public boolean supports(Class<?> type) {
-    return StringUtils.startsWithAny(
-        type.getPackageName(), reflectionToStringProperties.getBasePackages());
+    for (Class<?> supportedBaseClass : supportedBaseClasses) {
+      if (supportedBaseClass.isAssignableFrom(type)) {
+        return true;
+      }
+    }
+    return false;
   }
 
   @Override